fix(external-sync): fix root-skill sparse checkout, decompose git plumbing
Scoped slice of backlog 122 (items 2-3; item 1's skill_includes field
stays deferred -- speculative schema for a root skill that doesn't
exist yet, no current consumer).

Item 2: for skills_path: "." root-skill sources, set_sparse called
`git sparse-checkout disable`, which undoes the cone-mode
top-level-files-only default `git clone --sparse` already sets and
pulls the entire upstream working tree into the gitignored
_checkouts cache -- for herdr, the whole Rust app. Fix:
`sparse-checkout set` with no path args, which re-asserts
top-level-only explicitly and idempotently regardless of prior
sparse state on a reused cached checkout. Live-verified against the
real herdr checkout (not a synthetic fixture): re-running sync
against the existing already-fully-cloned 44M cache shrank it to
12M, working tree now shows only top-level files. Vendored output
confirmed byte-identical before/after.

Item 3: external_sync.rs mixed registry parsing, git plumbing, install,
and cleanup, sitting near the god-file ceiling. Extracted
ensure_checkout/set_sparse/resolve_ref_to_sha/checkout_sha and their
shared run_checked helper into a new external_sync_git.rs submodule,
same #[path = ...] convention the file already used for its own test
submodule. external_sync.rs dropped from 821 to 716 lines --
comfortably under the general 800-line ceiling, so its god-file
baseline entry was removed entirely rather than updated. All 8
existing tests pass unmodified.

Refs-backlog: 122

Agent: overnight autonomous lane
Agent-Surface: Claude Code
Agent-Model: claude-sonnet-5
