+++ crates/harness-kit-checks/baselines/godfiles.txt
@@ -3,7 +3,6 @@
 # Regenerate intentionally: harness-kit-checks check-godfiles --write-baseline
 859 crates/harness-kit-checks/src/config_loader.rs
 1883 crates/harness-kit-checks/src/docs_site.rs
-821 crates/harness-kit-checks/src/external_sync.rs
 1240 crates/harness-kit-checks/src/main.rs
 1014 crates/harness-kit-checks/src/skill_invocation_analytics.rs
 2465 crates/harness-kit-hooks/src/claude_hooks.rs
